+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Switch to absolute imports
 - A node's product can now be set from the outside
 - SupplyChain finds or creates nodes based on config
+- Update graph files on disk, separate nodes and edges in dot file
 
 ### Removed